==> code/10_subclustering/02-PCA_harmony_subcluster.Rmd <==
---
title: "02-PCA_harmony_subcluster"
author: "Bernie Mulvey"
date: "2025-02-06"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(fig.height = 10,fig.width = 7,include = FALSE)
#### sets tab autocompletion for directories to begin from the directory containing the .Rproj session, instead of the script's directory if different
knitr::opts_knit$set(root.dir = here::here())

library(data.table)
library(SpatialExperiment)
library(scater)
library(scran)
library(bluster)
library(Matrix)
library(harmony)

## rstudio GUI tweaks
require(colorout)
ColorOut()
# options("styler.addins_style_transformer" = "biocthis::bioc_style()")
##

## enable forked parallel processing with BiocParallel::multicoreParam, future::, etc. seemed to need this a couple times, but otherwise havent so its here as a preventative measure. part of this is adding the line 
# OBJC_DISABLE_INITIALIZE_FORK_SAFETY=YES
# to Renviron.site. see e.g. top response on https://stackoverflow.com/questions/73638290/python-on-mac-is-it-safe-to-set-objc-disable-initialize-fork-safety-yes-globall 
library(parallelly)
options(parallelly.supportsMulticore.disableOn="")
options(parallelly.fork.enable=TRUE)
library(BiocParallel)
options(bphost="localhost")
library(parallel)

## last of all, unload the base package datasets, whose data keep getting in the way of autocompletions (e.g., Theoph is priortized over TRUE)
unloadNamespace("datasets")
```

# repeated from script 01: load sectionwise, filtered, lognormed SPE; append the initial cluster labels (25HDG-75SVG louvain res 1)
```{r}
lc3 <- readRDS("processed-data/06-countsOnly_noImgData_QCed_SPE_split_to_tissSections.RDS")

louvs <- readRDS("processed-data/07_featureSelection_dimred_harmony_clustering/04-Initial_louvLeid_res0.5-1-2_clusterings.RDS")
louvs <- louvs <- louvs$HARMONYlmbna_HDG_SVG_2575[,.(rn,snnHARMONYlmbna_HDG_SVG_2575_louv_res1)]
setnames(louvs,2,"clusid")

autoan <- readRDS("processed-data/07_featureSelection_dimred_harmony_clustering/05b-init_clusterings_autoannoLCrapheAndGABAergic.RDS")
autoan <- autoan$snnHARMONYlmbna_HDG_SVG_2575_louv_res1

louvs <- merge.data.table(louvs,autoan,by="clusid",all.x=T)
louvs <- DataFrame(louvs,row.names=louvs$rn)[colnames(lc3),]

colLabels(lc3) <- louvs$autoanno
```

## repeated from script 01: make an SPE for each cluster individually
```{r}
clussub <- lapply(unique(colLabels(lc3)),function(x){ 
    sub <- lc3[,colLabels(lc3)==x]
    
    # drop genes that have 0 counts for the cluster in question
    sub <- sub[rowSums(counts(sub))>0,]

    # calculate the lognorm counts
    sub <- computeLibraryFactors(sub)
    sub <- logNormCounts(sub)
    return(sub)
})
names(clussub) <- unique(colLabels(lc3))

## we don't need the full spe now
rm(lc3,louvs,autoan)
gc(full=T)
```

## now load HVG sets for each cluster
```{r}
subhvgsets <- readRDS("processed-data/10_subclustering/01-HVG5-10-20_sets_per_cluster.RDS")
```

now for each cluster-feature set, run PCA, HARMONY and return the dimensionality reduction matrices

## WARNING: THIS IS SLOW AS HELL--dimensionality reduction steps need to be serialized for some reason. none of the parallelization packages seem to cooperate even though this all works with serial apply functions. ##

```{r}

## keep data.table from multithreading and thus overcommiting CPUs
setDTthreads(1,restore_after_fork = FALSE) # prevents reloading of data.table in default multithreaded mode

subdrs <- mapply(subclus=clussub,n=names(clussub),SIMPLIFY=FALSE,FUN=function(subclus,n){
    # get the HVG sets for the cluster
    hvgsets <- subhvgsets[[n]]
    
    dimrspes <- lapply(hvgsets,function(x){
        spe <- copy(subclus[x$gene_id,])
        return(spe)
    })
    names(dimrspes) <- paste0(n,"_",names(hvgsets)) # eg X1_hvg5
    
    # psockcl <- makeClusterPSOCK(c("localhost","localhost","localhost"))
    drtabs <- mapply(subspe=dimrspes,drn=names(dimrspes),SIMPLIFY=FALSE,FUN=function(subspe,drn){

        # standard PCAs
        set.seed(42)
        subspe <- scater::runPCA(subspe,name=paste0("PCA_",drn)) # eg PCA_X1_hvg5

        # standard UMAPs (which require PCA first)
        set.seed(42)
        subspe <- scater::runUMAP(subspe,dimred=paste0("PCA_",drn),name=paste0("UMAP_",drn))

        # harmony runs -- we need to temporarily rename the PCA for the feature set to "PCA" for runharmony to accept it
        reducedDimNames(subspe)[which(reducedDimNames(subspe)==paste0("PCA_",drn))] <- "PCA"
        
        # data-based initialization of lambda (development ver of Harmony; requires explicit lambda=NULL)
        set.seed(42)
        subspe <- RunHarmony(subspe,lambda=NULL,group.by.vars="sample_id",reduction.save=paste0("HARMONYlmbna_",drn))

        # restore the feature-set-specific name for the uncorrected PCA
        reducedDimNames(subspe)[which(reducedDimNames(subspe)=="PCA")] <- paste0("PCA_",drn)

        # UMAP of data-based initialization of lambda
        set.seed(42)
        subspe <- scater::runUMAP(subspe,dimred=paste0("HARMONYlmbna_",drn),name=paste0("HARMONYlmbna_",drn,"_UMAP"))

        ## grab the produced dimreds, drop the 10x-auto-dimreds if they're still in there, and return them in a list
        rdm <- reducedDims(subspe)
        rdm <- as.list(rdm)
        rdm <- rdm[grep(names(rdm),pattern="^10x",value=T,invert=T)]

        return(rdm)
    })
    
    names(drtabs) <- names(dimrspes)
    return(drtabs)
})
names(subdrs) <- names(clussub)
```

## save
```{r}
saveRDS(subdrs,"processed-data/10_subclustering/02a-subclustering_HVG5-10-20_PCA_harmony_umap_dimredmats.RDS")
```

now run SNN k10 --> louvain at resolutions seq(0.2,1,0.2) on each of the HARMONY dimreds. this part really truly should parallelize without issue and i will scream if it doesn't

```{r}
# clean up the workspace
rm(clussub,subhvgsets)
gc(full=T)

# simplify down to one big list of harmony dimensionality reductions
hdrs <- lapply(subdrs,function(s){
    sfg <- lapply(s,function(f){
        return(f[[3]])
    })
    names(sfg) <- unlist(lapply(s,function(f){names(f)[3]}))
    return(sfg)
})

## collapse down to one long list of harmony mats
hdrl <- list()
i<-1
for (i in c(1:length(hdrs))){
    hdrl <- c(hdrl,hdrs[[i]])
}
##

stopifnot(length(hdrl)==length(hdrs)*3)
stopifnot(length(unique(names(hdrl)))==length(hdrl))

## clean up
rm(i,hdrs,subdrs)
gc(full=T)
```

## now run the clustering
```{r}
multiressubclus <- mapply(dr=hdrl,drn=names(hdrl),SIMPLIFY=FALSE,FUN=function(dr,drn){
    snn <- makeSNNGraph(dr, k=10,BPPARAM=MulticoreParam(8))
    louvlist <- lapply(seq(0.2,1,0.2),function(res){
        set.seed(42)
        louv <- igraph::cluster_louvain(snn,resolution=res)
        louv <- paste0("X",louv$membership) # these come back in the same order as the rownames of the dimensionality reduction, which itself is in the same row order as the SPE is in columns. so we can just return the vector.
        return(louv)
    })
    names(louvlist) <- paste0(drn,"_louv",seq(0.2,1,0.2))
    
    
    
    rettab <- as.data.table(cbind(rownames(dr),as.data.frame(louvlist)))
    setnames(rettab,1,"rn")
    return(rettab)
})
names(multiressubclus) <- names(hdrl)

# save
saveRDS(multiressubclus,"processed-data/10_subclustering/02b-subclustering_HVG5-10-20_HARMONY_louv_res0.2-1.RDS")
```

rep inf
```{r}
sessionInfo()
sessioninfo::session_info()
```

R version 4.4.2 (2024-10-31)
Platform: aarch64-apple-darwin20
Running under: macOS Sonoma 14.7.1

Matrix products: default
BLAS:   /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib 
LAPACK: /Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/lib/libRlapack.dylib;  LAPACK version 3.12.0

locale:
[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8

time zone: America/Chicago
tzcode source: internal

attached base packages:
[1] parallel  stats4    stats     graphics  grDevices utils     methods   base     

other attached packages:
 [1] BiocParallel_1.40.0         parallelly_1.42.0          
 [3] colorout_1.3-0.2            harmony_1.2.3              
 [5] Rcpp_1.0.14                 Matrix_1.7-1               
 [7] bluster_1.16.0              scran_1.34.0               
 [9] scater_1.34.0               ggplot2_3.5.1              
[11] scuttle_1.16.0              SpatialExperiment_1.16.0   
[13] SingleCellExperiment_1.28.1 SummarizedExperiment_1.36.0
[15] Biobase_2.66.0              GenomicRanges_1.58.0       
[17] GenomeInfoDb_1.42.1         IRanges_2.40.1             
[19] S4Vectors_0.44.0            BiocGenerics_0.52.0        
[21] MatrixGenerics_1.18.0       matrixStats_1.5.0          
[23] data.table_1.16.4           rlang_1.1.4                

loaded via a namespace (and not attached):
 [1] tidyselect_1.2.1        viridisLite_0.4.2       dplyr_1.1.4            
 [4] vipor_0.4.7             viridis_0.6.5           fastmap_1.2.0          
 [7] digest_0.6.37           rsvd_1.0.5              lifecycle_1.0.4        
[10] cluster_2.1.6           statmod_1.5.0           magrittr_2.0.3         
[13] compiler_4.4.2          tools_4.4.2             igraph_2.1.4           
[16] yaml_2.3.10             knitr_1.49              S4Arrays_1.6.0         
[19] dqrng_0.4.1             here_1.0.1              DelayedArray_0.32.0    
[22] abind_1.4-8             withr_3.0.2             grid_4.4.2             
[25] beachmat_2.22.0         colorspace_2.1-1        edgeR_4.4.1            
[28] scales_1.3.0            cli_3.6.3               rmarkdown_2.29         
[31] crayon_1.5.3            generics_0.1.3          metapod_1.14.0         
[34] rstudioapi_0.17.1       httr_1.4.7              rjson_0.2.23           
[37] ggbeeswarm_0.7.2        zlibbioc_1.52.0         XVector_0.46.0         
[40] vctrs_0.6.5             jsonlite_1.8.9          BiocSingular_1.22.0    
[43] BiocNeighbors_2.0.1     ggrepel_0.9.6           irlba_2.3.5.1          
[46] beeswarm_0.4.0          magick_2.8.5            locfit_1.5-9.10        
[49] limma_3.62.1            glue_1.8.0              codetools_0.2-20       
[52] uwot_0.2.2              cowplot_1.1.3           RcppAnnoy_0.0.22       
[55] gtable_0.3.6            UCSC.utils_1.2.0        ScaledMatrix_1.14.0    
[58] munsell_0.5.1           tibble_3.2.1            pillar_1.10.1          
[61] htmltools_0.5.8.1       GenomeInfoDbData_1.2.13 R6_2.5.1               
[64] rprojroot_2.0.4         evaluate_1.0.3          lattice_0.22-6         
[67] RhpcBLASctl_0.23-42     gridExtra_2.3           SparseArray_1.6.0      
[70] xfun_0.50               pkgconfig_2.0.3        
> sessioninfo::session_info()
─ Session info ─────────────────────────────────────────────────────────────────────
 setting  value
 version  R version 4.4.2 (2024-10-31)
 os       macOS Sonoma 14.7.1
 system   aarch64, darwin20
 ui       RStudio
 language (EN)
 collate  en_US.UTF-8
 ctype    en_US.UTF-8
 tz       America/Chicago
 date     2025-02-06
 rstudio  2024.07.0-daily+219 Cranberry Hibiscus (desktop)
 pandoc   3.1.11 @ /Applications/RStudio.app/Contents/Resources/app/quarto/bin/tools/aarch64/ (via rmarkdown)

─ Packages ─────────────────────────────────────────────────────────────────────────
 !  package              * version  date (UTC) lib source
    abind                  1.4-8    2024-09-12 [1] CRAN (R 4.4.1)
    beachmat               2.22.0   2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    beeswarm               0.4.0    2021-06-01 [1] CRAN (R 4.4.0)
    Biobase              * 2.66.0   2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    BiocGenerics         * 0.52.0   2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    BiocNeighbors          2.0.1    2024-11-28 [1] Bioconductor 3.20 (R 4.4.2)
    BiocParallel         * 1.40.0   2024-11-23 [1] Bioconductor 3.20 (R 4.4.2)
    BiocSingular           1.22.0   2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    bluster              * 1.16.0   2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
 P  cli                    3.6.3    2024-06-21 [2] CRAN (R 4.4.0)
    cluster                2.1.6    2023-12-01 [1] CRAN (R 4.4.2)
    codetools              0.2-20   2024-03-31 [1] CRAN (R 4.4.2)
    colorout             * 1.3-0.2  2024-05-01 [1] Github (jalvesaq/colorout@c6113a2)
    colorspace             2.1-1    2024-07-26 [1] CRAN (R 4.4.0)
    cowplot                1.1.3    2024-01-22 [1] CRAN (R 4.4.0)
    crayon                 1.5.3    2024-06-20 [1] CRAN (R 4.4.0)
    data.table           * 1.16.4   2024-12-06 [1] CRAN (R 4.4.1)
    DelayedArray           0.32.0   2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    digest                 0.6.37   2024-08-19 [1] CRAN (R 4.4.1)
    dplyr                  1.1.4    2023-11-17 [1] CRAN (R 4.4.0)
    dqrng                  0.4.1    2024-05-28 [1] CRAN (R 4.4.0)
    edgeR                  4.4.1    2024-12-02 [1] Bioconductor 3.20 (R 4.4.2)
    evaluate               1.0.3    2025-01-10 [1] CRAN (R 4.4.1)
    fastmap                1.2.0    2024-05-15 [1] CRAN (R 4.4.0)
    generics               0.1.3    2022-07-05 [1] CRAN (R 4.4.0)
    GenomeInfoDb         * 1.42.1   2024-11-28 [1] Bioconductor 3.20 (R 4.4.2)
    GenomeInfoDbData       1.2.13   2024-12-12 [1] Bioconductor
    GenomicRanges        * 1.58.0   2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    ggbeeswarm             0.7.2    2023-04-29 [1] CRAN (R 4.4.0)
    ggplot2              * 3.5.1    2024-04-23 [1] CRAN (R 4.4.0)
    ggrepel                0.9.6    2024-09-07 [1] CRAN (R 4.4.1)
    glue                   1.8.0    2024-09-30 [1] CRAN (R 4.4.1)
    gridExtra              2.3      2017-09-09 [1] CRAN (R 4.4.0)
    gtable                 0.3.6    2024-10-25 [1] CRAN (R 4.4.1)
    harmony              * 1.2.3    2024-11-27 [1] CRAN (R 4.4.1)
    here                   1.0.1    2020-12-13 [1] CRAN (R 4.4.0)
    htmltools              0.5.8.1  2024-04-04 [1] CRAN (R 4.4.0)
    httr                   1.4.7    2023-08-15 [1] CRAN (R 4.4.0)
    igraph                 2.1.4    2025-01-23 [1] CRAN (R 4.4.1)
    IRanges              * 2.40.1   2024-12-05 [1] Bioconductor 3.20 (R 4.4.2)
    irlba                  2.3.5.1  2022-10-03 [1] CRAN (R 4.4.0)
    jsonlite               1.8.9    2024-09-20 [1] CRAN (R 4.4.1)
    knitr                  1.49     2024-11-08 [1] CRAN (R 4.4.1)
    lattice                0.22-6   2024-03-20 [1] CRAN (R 4.4.2)
    lifecycle              1.0.4    2023-11-07 [1] CRAN (R 4.4.0)
    limma                  3.62.1   2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    locfit                 1.5-9.10 2024-06-24 [1] CRAN (R 4.4.0)
    magick                 2.8.5    2024-09-20 [1] CRAN (R 4.4.1)
    magrittr               2.0.3    2022-03-30 [1] CRAN (R 4.4.0)
    Matrix               * 1.7-1    2024-10-18 [1] CRAN (R 4.4.2)
    MatrixGenerics       * 1.18.0   2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    matrixStats          * 1.5.0    2025-01-07 [1] CRAN (R 4.4.1)
    metapod                1.14.0   2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    munsell                0.5.1    2024-04-01 [1] CRAN (R 4.4.0)
    parallelly           * 1.42.0   2025-01-30 [1] CRAN (R 4.4.1)
    pillar                 1.10.1   2025-01-07 [1] CRAN (R 4.4.1)
    pkgconfig              2.0.3    2019-09-22 [1] CRAN (R 4.4.0)
    R6                     2.5.1    2021-08-19 [1] CRAN (R 4.4.0)
    Rcpp                 * 1.0.14   2025-01-12 [1] CRAN (R 4.4.1)
    RcppAnnoy              0.0.22   2024-01-23 [1] CRAN (R 4.4.0)
    RhpcBLASctl            0.23-42  2023-02-11 [1] CRAN (R 4.4.0)
    rjson                  0.2.23   2024-09-16 [1] CRAN (R 4.4.1)
 VP rlang                * 1.1.4    2025-01-17 [2] CRAN (R 4.4.1) (on disk 1.1.5)
    rmarkdown              2.29     2024-11-04 [1] CRAN (R 4.4.1)
    rprojroot              2.0.4    2023-11-05 [1] CRAN (R 4.4.0)
    rstudioapi             0.17.1   2024-10-22 [1] CRAN (R 4.4.1)
    rsvd                   1.0.5    2021-04-16 [1] CRAN (R 4.4.0)
    S4Arrays               1.6.0    2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    S4Vectors            * 0.44.0   2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    ScaledMatrix           1.14.0   2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    scales                 1.3.0    2023-11-28 [1] CRAN (R 4.4.0)
    scater               * 1.34.0   2024-10-29 [1] Bioconductor 3.20 (R 4.4.1)
    scran                * 1.34.0   2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    scuttle              * 1.16.0   2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    sessioninfo            1.2.2    2021-12-06 [1] CRAN (R 4.4.0)
    SingleCellExperiment * 1.28.1   2024-11-23 [1] Bioconductor 3.20 (R 4.4.2)
    SparseArray            1.6.0    2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    SpatialExperiment    * 1.16.0   2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    statmod                1.5.0    2023-01-06 [1] CRAN (R 4.4.0)
    SummarizedExperiment * 1.36.0   2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    tibble                 3.2.1    2023-03-20 [1] CRAN (R 4.4.0)
    tidyselect             1.2.1    2024-03-11 [1] CRAN (R 4.4.0)
    UCSC.utils             1.2.0    2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    uwot                   0.2.2    2024-04-21 [1] CRAN (R 4.4.0)
    vctrs                  0.6.5    2023-12-01 [1] CRAN (R 4.4.0)
    vipor                  0.4.7    2023-12-18 [1] CRAN (R 4.4.0)
    viridis                0.6.5    2024-01-29 [1] CRAN (R 4.4.0)
    viridisLite            0.4.2    2023-05-02 [1] CRAN (R 4.4.0)
    withr                  3.0.2    2024-10-28 [1] CRAN (R 4.4.1)
    xfun                   0.50     2025-01-07 [1] CRAN (R 4.4.1)
    XVector                0.46.0   2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    yaml                   2.3.10   2024-07-26 [1] CRAN (R 4.4.0)
    zlibbioc               1.52.0   2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)

 [1] /Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/library
 [2] /Users/bmulvey/Library/R/arm64/4.4/library

 V ── Loaded and on-disk version mismatch.
 P ── Loaded and on-disk path mismatch.

────────────────────────────────────────────────────────────────────────────────────
