==> code/10_subclustering/05-LCNM_vs_LCnonNM_PRELIM.Rmd <==
---
title: "05-LCNM_vs_LCnonNM"
author: "Bernie Mulvey"
date: "2025-05-12"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(fig.height = 10,fig.width = 7,include = FALSE)
#### sets tab autocompletion for directories to begin from the directory containing the .Rproj session, instead of the script's directory if different
knitr::opts_knit$set(root.dir = here::here())

library(data.table)
library(spatialLIBD)
# library(SpatialFeatureExperiment)

## rstudio GUI tweaks
require(colorout)
ColorOut()
options("styler.addins_style_transformer" = "biocthis::bioc_style()")
##

## enable forked parallel processing with BiocParallel::multicoreParam, future::, etc. seemed to need this a couple times, but otherwise havent so its here as a preventative measure. part of this is adding the line 
# OBJC_DISABLE_INITIALIZE_FORK_SAFETY=YES
# to Renviron.site. see e.g. top response on https://stackoverflow.com/questions/73638290/python-on-mac-is-it-safe-to-set-objc-disable-initialize-fork-safety-yes-globall 
library(parallelly)
options(parallelly.supportsMulticore.disableOn="")
options(parallelly.fork.enable=TRUE)
library(BiocParallel)
options(bphost="localhost")

## last of all, unload the base package datasets, whose data keep getting in the way of autocompletions (e.g., Theoph is priortized over TRUE)
unloadNamespace("datasets")
```

## load spe, clusters, and table of labels and NMseg subclusters
```{r}
lc3 <- readRDS("processed-data/06-countsOnly_noImgData_QCed_SPE_split_to_tissSections.RDS")

louvs <- readRDS("processed-data/07_featureSelection_dimred_harmony_clustering/04-Initial_louvLeid_res0.5-1-2_clusterings.RDS")
louvs <- louvs$HARMONYlmbna_HDG_SVG_2575[,.(rn,snnHARMONYlmbna_HDG_SVG_2575_louv_res1)]
setnames(louvs,2,"clusid")

anno <- fread("processed-data/08_validitycheck_25hdg75svg_louv1/10-25hdg75svg_louv1_annots.txt")
louvs <- merge.data.table(louvs,anno,by="clusid")
louvs <- DataFrame(louvs,row.names=louvs$rn)[colnames(lc3),]

colLabels(lc3) <- louvs$anno
rm(louvs,anno)

##
seglabs <- fread("processed-data/10_subclustering/04-LCsubclus-by-NM_otherclusts-by-LCorLCNMadjacency.txt")

## append these ^ and E4 copy number as a control variable
tmpcd <- as.data.table(colData(lc3),keep.rownames=T)
tmpcd <- merge.data.table(tmpcd,seglabs[,.(rn,LCadjacent,NMseg_subclus)],by="rn")

colData(lc3) <- DataFrame(tmpcd,row.names=tmpcd$rn)[colnames(lc3),]

## subset to LC cluster
lc3 <- lc3[,lc3$label=="LC"]
stopifnot(length(unique(lc3$NMseg_subclus))==2)

lc3$NMseg_subclus <- factor(lc3$NMseg_subclus,levels=c("LC_NMpos","LC_NMneg"))
```

## run marker analysis of NM+ vs NM- LC, controlling for age Sex and E4 copy number
```{r}
## fix APOE values to be limma passable
tmpcd <- as.data.table(colData(lc3))
tmpcd[APOE=="E2/ E3",APOE:="E2het"]
tmpcd[APOE=="E2/ E2",APOE:="E2hom"]
tmpcd[APOE=="E3/ E4",APOE:="E4het"]
tmpcd[APOE=="E4/ E4",APOE:="E4hom"]
colData(lc3) <- DataFrame(tmpcd,row.names=tmpcd$rn)[colnames(lc3),]

nmmk <- registration_wrapper(lc3,var_registration = "NMseg_subclus",var_sample_id = "sample_id",gene_ensembl = "gene_id",gene_name = "gene_name",min_ncells = 5,covars = c("Sex","Age","APOE"))

nmmk <- as.data.table(nmmk$enrichment)
keepn <- c("gene",grep(names(nmmk),pattern="NMpos",value=T),"ensembl")
nmmk <- nmmk[,..keepn]

nmmk <- melt.data.table(nmmk,id.vars=c("ensembl","gene"))
nmmk[,stat:=gsub(variable,pattern="^(.*)_LC.*$",replacement="\\1")]


nmmkt <- nmmk[stat=="t_stat"]
setnames(nmmkt,"value","t_stat")
nmmkt[,variable:=NULL]
nmmkt[,stat:=NULL]

nmmkp <- nmmk[stat=="p_value"]
setnames(nmmkp,"value","p_value")
nmmkp[,variable:=NULL]
nmmkp[,stat:=NULL]

nmmkf <- nmmk[stat=="fdr"]
setnames(nmmkf,"value","fdr")
nmmkf[,variable:=NULL]
nmmkf[,stat:=NULL]

nmmkl <- nmmk[stat=="logFC"]
setnames(nmmkl,"value","logFC")
nmmkl[,variable:=NULL]
nmmkl[,stat:=NULL]

nmmk2 <- merge.data.table(nmmkl,nmmkf,by=c("ensembl","gene"))
nmmk2 <- merge.data.table(nmmk2,nmmkp,by=c("ensembl","gene"))
nmmk2 <- merge.data.table(nmmk2,nmmkt,by=c("ensembl","gene"))

nmmk2[,clus:="LC_NMpos"]

nameord <- c("clus","gene","logFC","fdr","p_value","t_stat","ensembl")
nmmk2 <- nmmk2[,..nameord]
```

save
```{r}
fwrite(nmmk2,"processed-data/10_subclustering/05-LC_NMpos-vs-NMneg_markers_covarSexAgeE4copies.txt",sep='\t',quote=F,row.names=F,col.names=T)
```

repinf
```{r}
sessionInfo()
sessioninfo::session_info()
```
R version 4.4.3 (2025-02-28)
Platform: aarch64-apple-darwin20
Running under: macOS Sequoia 15.3.2

Matrix products: default
BLAS:   /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib 
LAPACK: /Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/lib/libRlapack.dylib;  LAPACK version 3.12.0

locale:
[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8

time zone: America/Chicago
tzcode source: internal

attached base packages:
[1] stats4    stats     graphics  grDevices utils     methods   base     

other attached packages:
 [1] BiocParallel_1.40.0         parallelly_1.43.0          
 [3] colorout_1.3-0.2            spatialLIBD_1.18.0         
 [5] SpatialExperiment_1.16.0    SingleCellExperiment_1.28.1
 [7] SummarizedExperiment_1.36.0 Biobase_2.66.0             
 [9] GenomicRanges_1.58.0        GenomeInfoDb_1.42.3        
[11] IRanges_2.40.1              S4Vectors_0.44.0           
[13] BiocGenerics_0.52.0         MatrixGenerics_1.18.1      
[15] matrixStats_1.5.0           data.table_1.17.0          
[17] rlang_1.1.5                

loaded via a namespace (and not attached):
  [1] RColorBrewer_1.1-3       rstudioapi_0.17.1        jsonlite_2.0.0          
  [4] magrittr_2.0.3           ggbeeswarm_0.7.2         magick_2.8.6            
  [7] farver_2.1.2             rmarkdown_2.29           BiocIO_1.16.0           
 [10] fields_16.3.1            zlibbioc_1.52.0          vctrs_0.6.5             
 [13] memoise_2.0.1            config_0.3.2             Rsamtools_2.22.0        
 [16] paletteer_1.6.0          RCurl_1.98-1.17          benchmarkme_1.0.8       
 [19] htmltools_0.5.8.1        S4Arrays_1.6.0           AnnotationHub_3.14.0    
 [22] curl_6.2.2               BiocNeighbors_2.0.1      SparseArray_1.6.2       
 [25] sass_0.4.10              bslib_0.9.0              htmlwidgets_1.6.4       
 [28] plotly_4.10.4            cachem_1.1.0             GenomicAlignments_1.42.0
 [31] mime_0.13                lifecycle_1.0.4          iterators_1.0.14        
 [34] pkgconfig_2.0.3          rsvd_1.0.5               Matrix_1.7-3            
 [37] R6_2.6.1                 fastmap_1.2.0            GenomeInfoDbData_1.2.13 
 [40] shiny_1.10.0             digest_0.6.37            rematch2_2.1.2          
 [43] AnnotationDbi_1.68.0     rprojroot_2.0.4          scater_1.34.1           
 [46] irlba_2.3.5.1            ExperimentHub_2.14.0     RSQLite_2.3.9           
 [49] beachmat_2.22.0          filelock_1.0.3           httr_1.4.7              
 [52] abind_1.4-8              compiler_4.4.3           here_1.0.1              
 [55] bit64_4.6.0-1            doParallel_1.0.17        attempt_0.3.1           
 [58] viridis_0.6.5            DBI_1.2.3                maps_3.4.2.1            
 [61] sessioninfo_1.2.3        rappdirs_0.3.3           DelayedArray_0.32.0     
 [64] rjson_0.2.23             tools_4.4.3              vipor_0.4.7             
 [67] beeswarm_0.4.0           httpuv_1.6.16            glue_1.8.0              
 [70] restfulr_0.0.15          promises_1.3.2           grid_4.4.3              
 [73] generics_0.1.3           gtable_0.3.6             tidyr_1.3.1             
 [76] ScaledMatrix_1.14.0      BiocSingular_1.22.0      XVector_0.46.0          
 [79] ggrepel_0.9.6            BiocVersion_3.20.0       foreach_1.5.2           
 [82] pillar_1.10.2            spam_2.11-1              limma_3.62.2            
 [85] later_1.4.2              benchmarkmeData_1.0.4    dplyr_1.1.4             
 [88] BiocFileCache_2.14.0     lattice_0.22-6           rtracklayer_1.66.0      
 [91] bit_4.6.0                tidyselect_1.2.1         locfit_1.5-9.12         
 [94] Biostrings_2.74.1        scuttle_1.16.0           knitr_1.50              
 [97] gridExtra_2.3            edgeR_4.4.2              xfun_0.52               
[100] statmod_1.5.0            DT_0.33                  UCSC.utils_1.2.0        
[103] lazyeval_0.2.2           yaml_2.3.10              shinyWidgets_0.9.0      
[106] evaluate_1.0.3           codetools_0.2-20         tibble_3.2.1            
[109] BiocManager_1.30.25      cli_3.6.4                xtable_1.8-4            
[112] jquerylib_0.1.4          golem_0.5.1              dichromat_2.0-0.1       
[115] Rcpp_1.0.14              dbplyr_2.5.0             png_0.1-8               
[118] XML_3.99-0.18            parallel_4.4.3           ggplot2_3.5.2           
[121] blob_1.2.4               dotCall64_1.2            bitops_1.0-9            
[124] viridisLite_0.4.2        scales_1.4.0             purrr_1.0.4             
[127] crayon_1.5.3             cowplot_1.1.3            KEGGREST_1.46.0         
> sessioninfo::session_info()
─ Session info ────────────────────────────────────────────────────────────────────
 setting  value
 version  R version 4.4.3 (2025-02-28)
 os       macOS Sequoia 15.3.2
 system   aarch64, darwin20
 ui       RStudio
 language (EN)
 collate  en_US.UTF-8
 ctype    en_US.UTF-8
 tz       America/Chicago
 date     2025-05-12
 rstudio  2024.07.0-daily+219 Cranberry Hibiscus (desktop)
 pandoc   3.1.11 @ /Applications/RStudio.app/Contents/Resources/app/quarto/bin/tools/aarch64/ (via rmarkdown)
 quarto   1.4.555 @ /Applications/RStudio.app/Contents/Resources/app/quarto/bin/quarto

─ Packages ────────────────────────────────────────────────────────────────────────
 !  package              * version   date (UTC) lib source
    abind                  1.4-8     2024-09-12 [1] CRAN (R 4.4.1)
    AnnotationDbi          1.68.0    2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    AnnotationHub          3.14.0    2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    attempt                0.3.1     2020-05-03 [1] CRAN (R 4.4.0)
    beachmat               2.22.0    2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    beeswarm               0.4.0     2021-06-01 [1] CRAN (R 4.4.0)
    benchmarkme            1.0.8     2022-06-12 [1] CRAN (R 4.4.0)
    benchmarkmeData        1.0.4     2020-04-23 [1] CRAN (R 4.4.0)
    Biobase              * 2.66.0    2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    BiocFileCache          2.14.0    2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    BiocGenerics         * 0.52.0    2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    BiocIO                 1.16.0    2024-12-12 [1] Github (Bioconductor/BiocIO@8e80624)
    BiocManager            1.30.25   2024-08-28 [1] CRAN (R 4.4.1)
    BiocNeighbors          2.0.1     2024-11-28 [1] Bioconductor 3.20 (R 4.4.2)
    BiocParallel         * 1.40.0    2024-11-23 [1] Bioconductor 3.20 (R 4.4.2)
    BiocSingular           1.22.0    2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    BiocVersion            3.20.0    2024-05-04 [1] Bioconductor 3.20 (R 4.4.0)
    Biostrings             2.74.1    2024-12-16 [1] Bioconductor 3.20 (R 4.4.2)
    bit                    4.6.0     2025-03-06 [1] CRAN (R 4.4.1)
    bit64                  4.6.0-1   2025-01-16 [1] CRAN (R 4.4.1)
    bitops                 1.0-9     2024-10-03 [1] CRAN (R 4.4.1)
    blob                   1.2.4     2023-03-17 [1] CRAN (R 4.4.0)
    bslib                  0.9.0     2025-01-30 [1] CRAN (R 4.4.1)
    cachem                 1.1.0     2024-05-16 [1] CRAN (R 4.4.0)
 VP cli                    3.6.4     2025-04-23 [2] CRAN (R 4.4.1) (on disk 3.6.5)
    codetools              0.2-20    2024-03-31 [1] CRAN (R 4.4.3)
    colorout             * 1.3-0.2   2024-05-01 [1] Github (jalvesaq/colorout@c6113a2)
    config                 0.3.2     2023-08-30 [1] CRAN (R 4.4.0)
    cowplot                1.1.3     2024-01-22 [1] CRAN (R 4.4.0)
    crayon                 1.5.3     2024-06-20 [1] CRAN (R 4.4.0)
    curl                   6.2.2     2025-03-24 [1] CRAN (R 4.4.1)
    data.table           * 1.17.0    2025-02-22 [1] CRAN (R 4.4.1)
    DBI                    1.2.3     2024-06-02 [1] CRAN (R 4.4.0)
    dbplyr                 2.5.0     2024-03-19 [1] CRAN (R 4.4.0)
    DelayedArray           0.32.0    2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    dichromat              2.0-0.1   2022-05-02 [1] CRAN (R 4.4.0)
    digest                 0.6.37    2024-08-19 [1] CRAN (R 4.4.1)
    doParallel             1.0.17    2022-02-07 [1] CRAN (R 4.4.1)
    dotCall64              1.2       2024-10-04 [1] CRAN (R 4.4.1)
    dplyr                  1.1.4     2023-11-17 [1] CRAN (R 4.4.0)
    DT                     0.33      2024-04-04 [1] CRAN (R 4.4.0)
    edgeR                  4.4.2     2025-01-27 [1] Bioconductor 3.20 (R 4.4.2)
    evaluate               1.0.3     2025-01-10 [1] CRAN (R 4.4.1)
    ExperimentHub          2.14.0    2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    farver                 2.1.2     2024-05-13 [1] CRAN (R 4.4.0)
    fastmap                1.2.0     2024-05-15 [1] CRAN (R 4.4.0)
    fields                 16.3.1    2025-03-08 [1] CRAN (R 4.4.1)
    filelock               1.0.3     2023-12-11 [1] CRAN (R 4.4.0)
    foreach                1.5.2     2022-02-02 [1] CRAN (R 4.4.0)
    generics               0.1.3     2022-07-05 [1] CRAN (R 4.4.0)
    GenomeInfoDb         * 1.42.3    2025-01-27 [1] Bioconductor 3.20 (R 4.4.2)
    GenomeInfoDbData       1.2.13    2024-12-12 [1] Bioconductor
    GenomicAlignments      1.42.0    2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    GenomicRanges        * 1.58.0    2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    ggbeeswarm             0.7.2     2023-04-29 [1] CRAN (R 4.4.0)
    ggplot2                3.5.2     2025-04-09 [1] CRAN (R 4.4.1)
    ggrepel                0.9.6     2024-09-07 [1] CRAN (R 4.4.1)
    glue                   1.8.0     2024-09-30 [1] CRAN (R 4.4.1)
    golem                  0.5.1     2024-08-27 [1] CRAN (R 4.4.1)
    gridExtra              2.3       2017-09-09 [1] CRAN (R 4.4.0)
    gtable                 0.3.6     2024-10-25 [1] CRAN (R 4.4.1)
    here                   1.0.1     2020-12-13 [1] CRAN (R 4.4.0)
    htmltools              0.5.8.1   2024-04-04 [1] CRAN (R 4.4.0)
    htmlwidgets            1.6.4     2023-12-06 [1] CRAN (R 4.4.0)
    httpuv                 1.6.16    2025-04-16 [1] CRAN (R 4.4.1)
    httr                   1.4.7     2023-08-15 [1] CRAN (R 4.4.0)
    IRanges              * 2.40.1    2024-12-05 [1] Bioconductor 3.20 (R 4.4.2)
    irlba                  2.3.5.1   2022-10-03 [1] CRAN (R 4.4.0)
    iterators              1.0.14    2022-02-05 [1] CRAN (R 4.4.0)
    jquerylib              0.1.4     2021-04-26 [1] CRAN (R 4.4.0)
    jsonlite               2.0.0     2025-03-27 [1] CRAN (R 4.4.1)
    KEGGREST               1.46.0    2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    knitr                  1.50      2025-03-16 [1] CRAN (R 4.4.2)
    later                  1.4.2     2025-04-08 [1] CRAN (R 4.4.1)
    lattice                0.22-6    2024-03-20 [1] CRAN (R 4.4.3)
    lazyeval               0.2.2     2019-03-15 [1] CRAN (R 4.4.0)
    lifecycle              1.0.4     2023-11-07 [1] CRAN (R 4.4.0)
    limma                  3.62.2    2025-01-09 [1] Bioconductor 3.20 (R 4.4.2)
    locfit                 1.5-9.12  2025-03-05 [1] CRAN (R 4.4.1)
    magick                 2.8.6     2025-03-23 [1] CRAN (R 4.4.1)
    magrittr               2.0.3     2022-03-30 [1] CRAN (R 4.4.0)
    maps                   3.4.2.1   2024-11-10 [1] CRAN (R 4.4.1)
    Matrix                 1.7-3     2025-03-11 [1] CRAN (R 4.4.1)
    MatrixGenerics       * 1.18.1    2025-01-09 [1] Bioconductor 3.20 (R 4.4.2)
    matrixStats          * 1.5.0     2025-01-07 [1] CRAN (R 4.4.1)
    memoise                2.0.1     2021-11-26 [1] CRAN (R 4.4.0)
    mime                   0.13      2025-03-17 [1] CRAN (R 4.4.1)
    paletteer              1.6.0     2024-01-21 [1] CRAN (R 4.4.0)
    parallelly           * 1.43.0    2025-03-24 [1] CRAN (R 4.4.1)
    pillar                 1.10.2    2025-04-05 [1] CRAN (R 4.4.1)
    pkgconfig              2.0.3     2019-09-22 [1] CRAN (R 4.4.0)
    plotly                 4.10.4    2024-01-13 [1] CRAN (R 4.4.0)
    png                    0.1-8     2022-11-29 [1] CRAN (R 4.4.0)
    promises               1.3.2     2024-11-28 [1] CRAN (R 4.4.1)
    purrr                  1.0.4     2025-02-05 [1] CRAN (R 4.4.1)
    R6                     2.6.1     2025-02-15 [1] CRAN (R 4.4.1)
    rappdirs               0.3.3     2021-01-31 [1] CRAN (R 4.4.0)
    RColorBrewer           1.1-3     2022-04-03 [1] CRAN (R 4.4.0)
    Rcpp                   1.0.14    2025-01-12 [1] CRAN (R 4.4.1)
    RCurl                  1.98-1.17 2025-03-22 [1] CRAN (R 4.4.1)
    rematch2               2.1.2     2020-05-01 [1] CRAN (R 4.4.0)
    restfulr               0.0.15    2022-06-16 [1] CRAN (R 4.4.0)
    rjson                  0.2.23    2024-09-16 [1] CRAN (R 4.4.1)
 VP rlang                * 1.1.5     2025-04-11 [2] CRAN (R 4.4.1) (on disk 1.1.6)
    rmarkdown              2.29      2024-11-04 [1] CRAN (R 4.4.1)
    rprojroot              2.0.4     2023-11-05 [1] CRAN (R 4.4.0)
    Rsamtools              2.22.0    2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    RSQLite                2.3.9     2024-12-03 [1] CRAN (R 4.4.1)
    rstudioapi             0.17.1    2024-10-22 [1] CRAN (R 4.4.1)
    rsvd                   1.0.5     2021-04-16 [1] CRAN (R 4.4.0)
    rtracklayer            1.66.0    2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    S4Arrays               1.6.0     2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    S4Vectors            * 0.44.0    2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    sass                   0.4.10    2025-04-11 [1] CRAN (R 4.4.1)
    ScaledMatrix           1.14.0    2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    scales                 1.4.0     2025-04-24 [1] CRAN (R 4.4.1)
    scater                 1.34.1    2025-03-03 [1] Bioconductor 3.20 (R 4.4.2)
    scuttle                1.16.0    2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    sessioninfo            1.2.3     2025-02-05 [1] CRAN (R 4.4.1)
    shiny                  1.10.0    2024-12-14 [1] CRAN (R 4.4.1)
    shinyWidgets           0.9.0     2025-02-21 [1] CRAN (R 4.4.1)
    SingleCellExperiment * 1.28.1    2024-11-23 [1] Bioconductor 3.20 (R 4.4.2)
    spam                   2.11-1    2025-01-20 [1] CRAN (R 4.4.1)
    SparseArray            1.6.2     2025-02-20 [1] Bioconductor 3.20 (R 4.4.2)
    SpatialExperiment    * 1.16.0    2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    spatialLIBD          * 1.18.0    2024-11-07 [1] Bioconductor 3.20 (R 4.4.3)
    statmod                1.5.0     2023-01-06 [1] CRAN (R 4.4.0)
    SummarizedExperiment * 1.36.0    2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    tibble                 3.2.1     2023-03-20 [1] CRAN (R 4.4.0)
    tidyr                  1.3.1     2024-01-24 [1] CRAN (R 4.4.0)
    tidyselect             1.2.1     2024-03-11 [1] CRAN (R 4.4.0)
    UCSC.utils             1.2.0     2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    vctrs                  0.6.5     2023-12-01 [1] CRAN (R 4.4.0)
    vipor                  0.4.7     2023-12-18 [1] CRAN (R 4.4.0)
    viridis                0.6.5     2024-01-29 [1] CRAN (R 4.4.0)
    viridisLite            0.4.2     2023-05-02 [1] CRAN (R 4.4.0)
    xfun                   0.52      2025-04-02 [1] CRAN (R 4.4.1)
    XML                    3.99-0.18 2025-01-01 [1] CRAN (R 4.4.1)
    xtable                 1.8-4     2019-04-21 [1] CRAN (R 4.4.0)
    XVector                0.46.0    2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)
    yaml                   2.3.10    2024-07-26 [1] CRAN (R 4.4.0)
    zlibbioc               1.52.0    2024-11-08 [1] Bioconductor 3.20 (R 4.4.1)

 [1] /Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/library
 [2] /Users/bmulvey/Library/R/arm64/4.4/library

 * ── Packages attached to the search path.
 V ── Loaded and on-disk version mismatch.
 P ── Loaded and on-disk path mismatch.

───────────────────────────────────────────────────────────────────────────────────
